mister-ottani-vendas/backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
psycopg2-binary
python-dotenv
# Adicione outras dependências conforme precisar
